20230808-hugo-github-pages.md: Add subtitle, modification date and update copyright
I finally figured out how to add a custom footer to add the subtitle.

Originally I wanted to add the subtitle below the title, but it simply
didn't look good, so I added it to the footer.

Removed the author box, which only showed my name at the at the bottom
of each post and updated the copyright line instead.